refacto: rename extensions to ext
